back/requirements.txt
aniso8601==9.0.1
astroid==2.15.6
attrs==23.1.0
bandit==1.7.5
blinker==1.6.2
cfgv==3.3.1
click==8.1.4
dill==0.3.6
distlib==0.3.6
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-JWT-Extended==4.5.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
identify==2.5.24
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
lazy-object-proxy==1.9.0
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
nodeenv==1.8.0
pbr==5.11.1
platformdirs==3.8.1
pre-commit==3.3.3
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rich==13.4.2
rpds-py==0.8.10
smmap==5.0.0
SQLAlchemy==2.0.18
stevedore==5.1.0
tomlkit==0.11.8
typing_extensions==4.7.1
virtualenv==20.23.1
Werkzeug==2.3.6
wrapt==1.15.0
